minor UI changes

+++ Books/routes.py
@@ -35,7 +35,7 @@ def register():
         db.session.add(member)
         db.session.commit()
         flash(f'member added succesfully {form.username.data}!','success')
-        return redirect(url_for('home'))
+        return redirect(url_for('Members'))
     return render_template('register.html', title='Register',form = form)
 
 @app.route("/Issue_Book",methods = ['GET','POST'])
@@ -55,16 +55,16 @@ def Issue_Book():
                 flash(f'{book.book_name} Issued successfully for {member.name}','success')
                 return redirect(url_for('home'))
             else:
-                flash(f'{book.book_name} is not in stock, kindly wait or issue another book!!!!','failure')
+                flash(f'{book.book_name} is not in stock, kindly wait or issue another book!!!!','danger')
         else:
-            flash(f'Please enter valid details!!!!!','failure')
+            flash(f'Please enter valid details!!!!!','danger')
     return render_template('Issue_Book.html',title = 'Issue Book',issue_data = issue_data)  
 
-@app.route("/Return_Book",methods=['GET','POST'])
-def Return_Book():
+@app.route("/Return_Book/<int:transactionId>",methods=['GET','POST'])
+def Return_Book(transactionId):
     return_data = Return_Book_Form()
     if return_data.validate_on_submit():
-        transaction= models.Transaction.query.filter_by(transaction_id = return_data.transaction_id.data).first()
+        transaction= models.Transaction.query.filter_by(transaction_id = transactionId).first()
         member = models.Member.query.filter_by(memberid = transaction.member_id).first()
         book = models.Book.query.filter_by(id = transaction.book_id).first()
         days = (date.today() - transaction.issue_date).days
@@ -74,7 +74,7 @@ def Return_Book():
             total_charge = days * transaction.per_day_fees
         outstanding_debt = member.debt
         if outstanding_debt - return_data.amount_paid.data + total_charge > 500:
-            flash(f'Your current debt plus total charge exceeds 500 INR value, kindly pay an amount greater than or equal to {outstanding_debt + total_charge - 500}','failure')
+            flash(f'Your current debt plus total charge exceeds 500 INR value, kindly pay an amount greater than or equal to {outstanding_debt + total_charge - 500}','danger')
             return redirect(url_for('Return_Book'))
         else:
             member.debt = member.debt - return_data.amount_paid.data + total_charge
@@ -111,7 +111,7 @@ def Increase_Quantity():
             flash(f'{book_information.quantity.data} added to {book.book_name}','success')
             return redirect(url_for('library'))
         else:
-            flash(f'Please enter a valid book ID!!!!!','failure')
+            flash(f'Please enter a valid book ID!!!!!','danger')
             return redirect(url_for('Increase_Quantity'))
     return render_template('Increase_Quantity.html',title = 'Increase Quantity',book_information = book_information)
 
@@ -120,14 +120,26 @@ def Delete_Books():
     book_information = DeleteBook()
     if book_information.validate_on_submit():
         book = models.Book.query.filter_by(id = book_information.book_id.data).first()
-        book_name = book.book_name
         if book != None:
-            db.session.delete(book)
-            db.session.commit()
-            flash(f'{book_name} Deleted from Library','success')
-            return redirect(url_for('library'))
+            book_name = book.book_name
+            book_id = book.id
+            transaction = models.Transaction.query.filter_by(book_id = book_id).first()
+            if transaction:
+                print(transaction.book_status)
+                if transaction.book_status == True:
+                    db.session.delete(book)
+                    db.session.commit()
+                    return redirect(url_for('library'))
+                else:
+                    flash(f'This book is currently issued by {transaction.member_name}, kindly wait for them to return the book','danger')
+                    return redirect(url_for('Delete_Books'))
+            else:
+                db.session.delete(book)
+                db.session.commit()
+                flash(f'{book_name} Deleted from Library','success')
+                return redirect(url_for('library'))
         else:
-            flash(f'Please enter a valid book ID!!!!!','failure')
+            flash(f'Please enter a valid book ID!!!!!','danger')
             return redirect(url_for('Delete_Books'))
     return render_template('Delete_Books.html',title = 'Delete Books',book_information = book_information)
 
@@ -136,14 +148,14 @@ def Delete_Member():
     member_info = DeleteMember()
     if member_info.validate_on_submit():
         member = models.Member.query.filter_by(memberid = member_info.member_id.data).first()
-        member_name = member.name
         if member != None:
+            member_name = member.name
             db.session.delete(member)
             db.session.commit()
             flash(f'{member_name} Removed from Library','success')
             return redirect(url_for('Members'))
         else:
-            flash(f'Please enter a valid member ID!!!!!','failure')
+            flash(f'Please enter a valid member ID!!!!!','danger')
             return redirect(url_for('Delete_Member'))
     return render_template('Delete_Member.html',title = 'Delete Member',member_info = member_info
     )
